src/requirements.txt
alembic==1.7.7
asyncpg==0.26.0
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
click==7.1.2
enum-tools==0.9.0.post1
google-api-core==2.8.2
google-auth==2.10.0
google-cloud-core==2.3.1
google-cloud-storage==2.0.0
google-cloud-vision==2.7.3
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
minio==7.1.11
numpy==1.19.5
proto-plus==1.22.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pytest-simplehttpserver==0.1.2
pytz==2022.2.1
requests==2.27.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
typing_extensions==4.1.1
urllib3==1.26.11
urlparser==0.1.2
zipp==3.6.0
